move session react components to a separate directory
