+++ ui/desktop/src/components/sessions/SessionListView.tsx
@@ -1,11 +1,11 @@
 import React, { useEffect, useState } from 'react';
-import { ViewConfig } from '../App';
+import { ViewConfig } from '../../App';
 import { MessageSquare, Loader, AlertCircle, Calendar, ChevronRight } from 'lucide-react';
-import { fetchSessions, type Session } from '../sessions';
-import { Card } from './ui/card';
-import { Button } from './ui/button';
-import BackButton from './ui/BackButton';
-import { ScrollArea } from './ui/scroll-area';
+import { fetchSessions, type Session } from '../../sessions';
+import { Card } from '../ui/card';
+import { Button } from '../ui/button';
+import BackButton from '../ui/BackButton';
+import { ScrollArea } from '../ui/scroll-area';
 
 interface SessionListViewProps {
   setView: (view: ViewConfig['view'], viewOptions?: Record<any, any>) => void;
